refactor: update navigation and remove unused components in venue profile
